requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.10.3
python-multipart>=0.0.9
httpx>=0.27.0
markitdown>=0.1.0
lightrag-hku[api]>=0.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
